pyproject.toml
[tool.poetry]
name = "repl_python3_APAC-McDelivery-Logger"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
bs4 = "^0.0.1"
pandas = "^1.3.5"
selenium = "^4.1.5"
Scrapy = "^2.6.1"
"ruamel.yaml" = "^0.17.21"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
